Add SLURM interface (#96)
* Add validation for `use_slurm` flag in input.go

* Add job file creation and sbatch function

This commit adds the functionality to create a job file and run sbatch. It includes the implementation of the `CreateJobHeader`, `CreateJobBody`, `PrepareJobFile`, and `Sbatch` functions. Tests for these functions have also been added.

* Refactor job file creation and remove unnecessary test cases

* Add support for Slurm job submission

This commit adds support for submitting jobs using Slurm. It checks if there is a `job.sh` file in the run directory and if so, it uses `sbatch` to submit the job. Otherwise, it continues to use the existing command to run the job. Additionally, a new method `PrepareJobFile` is added to create the `job.sh` file.

* upgrade trunk

* Update execution mode key in input validation

* update dev container setup

* update main.go

* Refactor Job struct and add support for detecting submitted jobs

* Add FindNewestLogFile function to utils.go and corresponding unit test

This commit adds a new function called FindNewestLogFile to the utils.go file. This function finds the newest log file in a given directory by comparing the modification times of all the files in the directory. It also includes a corresponding unit test in the utils_test.go file to verify the functionality of the new function.

* Add logging for job status and elapsed time